feat(recurring-swap): add OrdersPickerScreen + extract shared DrumColumn (CP-13663)
