+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/swap/_layout.tsx
@@ -32,6 +32,7 @@ export default function SwapLayout(): JSX.Element {
           <Stack.Screen name="pricingDetails" />
           <Stack.Screen name="slippageDetails" />
           <Stack.Screen name="recurring/frequency" />
+          <Stack.Screen name="recurring/orders" />
         </Stack>
       </RecurringSwapContextProvider>
     </SwapContextProvider>
