+++ packages/core-mobile/app/new/routes/(signedIn)/(modals)/swap/recurring/orders.tsx
@@ -1,0 +1,3 @@
+import { OrdersPickerScreen } from 'features/recurringSwap/screens/OrdersPickerScreen'
+
+export default OrdersPickerScreen
